+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `step_embed()` now correctly defaults to have a random id with the word "embed". (#102)
 
+* `step_feature_hash()` is soft deprecated in embed in favor of `step_dummy_hash()` in textrecipes. (#95)
+
 # embed 0.1.5
 
 * Re-licensed package from GPL-2 to MIT. See [consent from copyright holders here](https://github.com/tidymodels/embed/issues/78).